e2e: Refactor suite setup and helpers to tests/fixture/e2e for reuse by coreth (#2265)
