+++ tests/e2e/e2e_test.go
@@ -4,22 +4,13 @@
 package e2e_test
 
 import (
-	"encoding/json"
-	"flag"
-	"fmt"
-	"os"
 	"testing"
 
 	ginkgo "github.com/onsi/ginkgo/v2"
 
 	"github.com/onsi/gomega"
 
-	"github.com/stretchr/testify/require"
-
-	"github.com/ava-labs/avalanchego/tests"
-	"github.com/ava-labs/avalanchego/tests/e2e"
-	"github.com/ava-labs/avalanchego/tests/fixture"
-	"github.com/ava-labs/avalanchego/tests/fixture/testnet/local"
+	"github.com/ava-labs/avalanchego/tests/fixture/e2e"
 
 	// ensure test packages are scanned by ginkgo
 	_ "github.com/ava-labs/avalanchego/tests/e2e/banff"
@@ -36,75 +27,18 @@ func TestE2E(t *testing.T) {
 	ginkgo.RunSpecs(t, "e2e test suites")
 }
 
-var (
-	avalancheGoExecPath  string
-	persistentNetworkDir string
-	usePersistentNetwork bool
-)
+var flagVars *e2e.FlagVars
 
 func init() {
-	flag.StringVar(
-		&avalancheGoExecPath,
-		"avalanchego-path",
-		os.Getenv(local.AvalancheGoPathEnvName),
-		fmt.Sprintf("avalanchego executable path (required if not using a persistent network). Also possible to configure via the %s env variable.", local.AvalancheGoPathEnvName),
-	)
-	flag.StringVar(
-		&persistentNetworkDir,
-		"network-dir",
-		"",
-		fmt.Sprintf("[optional] the dir containing the configuration of a persistent network to target for testing. Useful for speeding up test development. Also possible to configure via the %s env variable.", local.NetworkDirEnvName),
-	)
-	flag.BoolVar(
-		&usePersistentNetwork,
-		"use-persistent-network",
-		false,
-		"[optional] whether to target the persistent network identified by --network-dir.",
-	)
+	flagVars = e2e.RegisterFlags()
 }
 
 var _ = ginkgo.SynchronizedBeforeSuite(func() []byte {
 	// Run only once in the first ginkgo process
-
-	require := require.New(ginkgo.GinkgoT())
-
-	if usePersistentNetwork && len(persistentNetworkDir) == 0 {
-		persistentNetworkDir = os.Getenv(local.NetworkDirEnvName)
-	}
-
-	// Load or create a test network
-	var network *local.LocalNetwork
-	if len(persistentNetworkDir) > 0 {
-		tests.Outf("{{yellow}}Using a persistent network configured at %s{{/}}\n", persistentNetworkDir)
-
-		var err error
-		network, err = local.ReadNetwork(persistentNetworkDir)
-		require.NoError(err)
-	} else {
-		network = e2e.StartLocalNetwork(avalancheGoExecPath, e2e.DefaultNetworkDir)
-	}
-
-	uris := network.GetURIs()
-	require.NotEmpty(uris, "network contains no nodes")
-	tests.Outf("{{green}}network URIs: {{/}} %+v\n", uris)
-
-	testDataServerURI, err := fixture.ServeTestData(fixture.TestData{
-		FundedKeys: network.FundedKeys,
-	})
-	tests.Outf("{{green}}test data server URI: {{/}} %+v\n", testDataServerURI)
-	require.NoError(err)
-
-	env := &e2e.TestEnvironment{
-		NetworkDir:        network.Dir,
-		URIs:              uris,
-		TestDataServerURI: testDataServerURI,
-	}
-	bytes, err := json.Marshal(env)
-	require.NoError(err)
-	return bytes
+	return e2e.NewTestEnvironment(flagVars).Marshal()
 }, func(envBytes []byte) {
 	// Run in every ginkgo process
 
 	// Initialize the local test environment from the global state
-	e2e.InitTestEnvironment(envBytes)
+	e2e.InitSharedTestEnvironment(envBytes)
 })
