+++ tests/fixture/e2e/helpers.go
@@ -1,18 +1,14 @@
 // Copyright (C) 2019-2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-// e2e implements the e2e tests.
 package e2e
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"math/big"
-	"math/rand"
 	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
@@ -26,11 +22,8 @@ import (
 
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/tests"
-	"github.com/ava-labs/avalanchego/tests/fixture"
 	"github.com/ava-labs/avalanchego/tests/fixture/testnet"
 	"github.com/ava-labs/avalanchego/tests/fixture/testnet/local"
-	"github.com/ava-labs/avalanchego/utils/crypto/secp256k1"
-	"github.com/ava-labs/avalanchego/utils/perms"
 	"github.com/ava-labs/avalanchego/vms/platformvm/txs/executor"
 	"github.com/ava-labs/avalanchego/vms/secp256k1fx"
 	"github.com/ava-labs/avalanchego/wallet/subnet/primary"
@@ -67,79 +60,6 @@ const (
 	// under the shared network dir.
 	PrivateNetworksDirName = "private_networks"
 )
-
-// Env is used to access shared test fixture. Intended to be
-// initialized by SynchronizedBeforeSuite.
-var Env *TestEnvironment
-
-type TestEnvironment struct {
-	// The directory where the test network configuration is stored
-	NetworkDir string
-	// URIs used to access the API endpoints of nodes of the network
-	URIs []testnet.NodeURI
-	// The URI used to access the http server that allocates test data
-	TestDataServerURI string
-
-	require *require.Assertions
-}
-
-func InitTestEnvironment(envBytes []byte) {
-	require := require.New(ginkgo.GinkgoT())
-	require.Nil(Env, "env already initialized")
-	Env = &TestEnvironment{
-		require: require,
-	}
-	require.NoError(json.Unmarshal(envBytes, Env))
-}
-
-// Retrieve a random URI to naively attempt to spread API load across
-// nodes.
-func (te *TestEnvironment) GetRandomNodeURI() testnet.NodeURI {
-	r := rand.New(rand.NewSource(time.Now().Unix())) //#nosec G404
-	nodeURI := te.URIs[r.Intn(len(te.URIs))]
-	tests.Outf("{{blue}} targeting node %s with URI: %s{{/}}\n", nodeURI.NodeID, nodeURI.URI)
-	return nodeURI
-}
-
-// Retrieve the network to target for testing.
-func (te *TestEnvironment) GetNetwork() testnet.Network {
-	network, err := local.ReadNetwork(te.NetworkDir)
-	te.require.NoError(err)
-	return network
-}
-
-// Retrieve the specified number of funded keys allocated for the caller's exclusive use.
-func (te *TestEnvironment) AllocateFundedKeys(count int) []*secp256k1.PrivateKey {
-	keys, err := fixture.AllocateFundedKeys(te.TestDataServerURI, count)
-	te.require.NoError(err)
-	tests.Outf("{{blue}} allocated funded key(s): %+v{{/}}\n", keys)
-	return keys
-}
-
-// Retrieve a funded key allocated for the caller's exclusive use.
-func (te *TestEnvironment) AllocateFundedKey() *secp256k1.PrivateKey {
-	return te.AllocateFundedKeys(1)[0]
-}
-
-// Create a new keychain with the specified number of test keys.
-func (te *TestEnvironment) NewKeychain(count int) *secp256k1fx.Keychain {
-	keys := te.AllocateFundedKeys(count)
-	return secp256k1fx.NewKeychain(keys...)
-}
-
-// Create a new private network that is not shared with other tests.
-func (te *TestEnvironment) NewPrivateNetwork() testnet.Network {
-	// Load the shared network to retrieve its path and exec path
-	sharedNetwork, err := local.ReadNetwork(te.NetworkDir)
-	te.require.NoError(err)
-
-	// The private networks dir is under the shared network dir to ensure it
-	// will be included in the artifact uploaded in CI.
-	privateNetworksDir := filepath.Join(sharedNetwork.Dir, PrivateNetworksDirName)
-	te.require.NoError(os.MkdirAll(privateNetworksDir, perms.ReadWriteExecute))
-
-	return StartLocalNetwork(sharedNetwork.ExecPath, privateNetworksDir)
-}
 
 // Create a new wallet for the provided keychain against the specified node URI.
 func NewWallet(keychain *secp256k1fx.Keychain, nodeURI testnet.NodeURI) primary.Wallet {
